Updated EDM unit testing so that it does not fail the ECC build process if a metrics database is not set up.
Removed the test module from the EDM pom file, so the tests are not automatically run. Also added an edm.properties file and a README for instructions on how to run the EDM tests. Updated unit tests so that each checks fails gracefully if a DB connection has not been made, instead of giving errors.