Routed entity enablement & 'live' enablement in ECC dashboard
Clients can now properly send 'enable' or 'disable' messages to the ECC
to enable or disable an entity from pull requests. The EM API now also allows
the experimenter to turn entity 'liveness' on or off, such that data will be
ignored from a client, if required. Dashboard UI support for this has not been
implemented yet.